+++ cmd/tempo-serverless/handler.go
@@ -27,7 +27,9 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-const envConfigPrefix = "TEMPO"
+const (
+	envConfigPrefix = "TEMPO"
+)
 
 // used to initialize a reader one time
 var (
@@ -48,6 +50,11 @@ func Handler(r *http.Request) (*tempopb.SearchResponse, *HTTPError) {
 	searchReq, err := api.ParseSearchBlockRequest(r)
 	if err != nil {
 		return nil, httpError("parsing search request", err, http.StatusBadRequest)
+	}
+
+	maxBytes, err := api.ExtractServerlessParams(r)
+	if err != nil {
+		return nil, httpError("extracting serverless params", err, http.StatusBadRequest)
 	}
 
 	// load config, fields are set through env vars TEMPO_
@@ -115,6 +122,11 @@ func Handler(r *http.Request) (*tempopb.SearchResponse, *HTTPError) {
 
 		resp.Metrics.InspectedTraces++
 		resp.Metrics.InspectedBytes += uint64(len(obj))
+
+		if maxBytes > 0 && len(obj) > maxBytes {
+			resp.Metrics.SkippedTraces++
+			continue
+		}
 
 		metadata, err := decoder.Matches(id, obj, searchReq.SearchReq)
 		if err != nil {