+++ tempodb/tempodb_test.go
@@ -86,7 +86,7 @@ func TestDB(t *testing.T) {
 
 		bReq, err := proto.Marshal(req)
 		assert.NoError(t, err)
-		err = head.Append(id, bReq)
+		err = head.Append(id, bReq, 0, 0)
 		assert.NoError(t, err, "unexpected error writing req")
 	}
 
@@ -133,7 +133,7 @@ func TestBlockSharding(t *testing.T) {
 
 	bReq, err := proto.Marshal(req)
 	assert.NoError(t, err)
-	err = head.Append(id, bReq)
+	err = head.Append(id, bReq, 0, 0)
 	assert.NoError(t, err, "unexpected error writing req")
 
 	// write block to backend
@@ -497,7 +497,7 @@ func TestSearchCompactedBlocks(t *testing.T) {
 
 		bReq, err := proto.Marshal(req)
 		assert.NoError(t, err)
-		err = head.Append(id, bReq)
+		err = head.Append(id, bReq, 0, 0)
 		assert.NoError(t, err, "unexpected error writing req")
 	}
 
@@ -575,7 +575,7 @@ func TestCompleteBlock(t *testing.T) {
 		ids = append(ids, id)
 		bReq, err := proto.Marshal(req)
 		assert.NoError(t, err)
-		err = block.Append(id, bReq)
+		err = block.Append(id, bReq, 0, 0)
 		assert.NoError(t, err, "unexpected error writing req")
 	}