+++ tempodb/wal/wal.go
@@ -16,6 +16,10 @@ import (
 	versioned_encoding "github.com/grafana/tempo/tempodb/encoding"
 )
 
+// extracts a time range from an object. start/end times returned are unix epoch
+// seconds
+type RangeFunc func(obj []byte, dataEncoding string) (uint32, uint32, error)
+
 const (
 	completedDir = "completed"
 	blocksDir    = "blocks"
@@ -80,7 +84,7 @@ func New(c *Config) (*WAL, error) {
 }
 
 // RescanBlocks returns a slice of append blocks from the wal folder
-func (w *WAL) RescanBlocks(log log.Logger) ([]*AppendBlock, error) {
+func (w *WAL) RescanBlocks(fn RangeFunc, log log.Logger) ([]*AppendBlock, error) {
 	files, err := os.ReadDir(w.c.Filepath)
 	if err != nil {
 		return nil, err
@@ -99,7 +103,7 @@ func (w *WAL) RescanBlocks(log log.Logger) ([]*AppendBlock, error) {
 		}
 
 		level.Info(log).Log("msg", "beginning replay", "file", f.Name(), "size", fileInfo.Size())
-		b, warning, err := newAppendBlockFromFile(f.Name(), w.c.Filepath)
+		b, warning, err := newAppendBlockFromFile(f.Name(), w.c.Filepath, fn)
 
 		remove := false
 		if err != nil {