+++ tempodb/wal/wal_test.go
@@ -9,6 +9,7 @@ import (
 	"path"
 	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/go-kit/log"
 	"github.com/golang/protobuf/proto"
@@ -58,7 +59,7 @@ func TestAppend(t *testing.T) {
 		reqs = append(reqs, req)
 		bReq, err := proto.Marshal(req)
 		require.NoError(t, err)
-		err = block.Append([]byte{0x01}, bReq)
+		err = block.Append([]byte{0x01}, bReq, 0, 0)
 		require.NoError(t, err, "unexpected error writing req")
 	}
 
@@ -132,7 +133,7 @@ func TestErrorConditions(t *testing.T) {
 		bObj, err := proto.Marshal(obj)
 		require.NoError(t, err)
 
-		err = block.Append(id, bObj)
+		err = block.Append(id, bObj, 0, 0)
 		require.NoError(t, err, "unexpected error writing req")
 	}
 	appendFile, err := os.OpenFile(block.fullFilename(), os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0600)
@@ -150,7 +151,9 @@ func TestErrorConditions(t *testing.T) {
 	err = os.WriteFile(filepath.Join(tempDir, "fe0b83eb-a86b-4b6c-9a74-dc272cd5700e:blerg:v2:gzip"), []byte{}, 0644)
 	require.NoError(t, err)
 
-	blocks, err := wal.RescanBlocks(log.NewNopLogger())
+	blocks, err := wal.RescanBlocks(func([]byte, string) (uint32, uint32, error) {
+		return 0, 0, nil
+	}, log.NewNopLogger())
 	require.NoError(t, err, "unexpected error getting blocks")
 	require.Len(t, blocks, 1)
 
@@ -159,6 +162,46 @@ func TestErrorConditions(t *testing.T) {
 	// confirm block has been removed
 	require.NoFileExists(t, filepath.Join(tempDir, "fe0b83eb-a86b-4b6c-9a74-dc272cd5700e:tenant:v2:gzip"))
 	require.NoFileExists(t, filepath.Join(tempDir, "fe0b83eb-a86b-4b6c-9a74-dc272cd5700e:blerg:v2:gzip"))
+}
+
+func TestAppendBlockStartEnd(t *testing.T) {
+	wal, err := New(&Config{
+		Filepath: t.TempDir(),
+		Encoding: backend.EncNone,
+	})
+	require.NoError(t, err, "unexpected error creating temp wal")
+
+	blockID := uuid.New()
+	block, err := wal.NewBlock(blockID, testTenantID, "")
+	require.NoError(t, err, "unexpected error creating block")
+
+	// create a new block and confirm start/end times are correct
+	blockStart := uint32(time.Now().Unix())
+	blockEnd := uint32(time.Now().Add(time.Minute).Unix())
+
+	for i := 0; i < 10; i++ {
+		bytes := make([]byte, 16)
+		rand.Read(bytes)
+
+		err = block.Append(bytes, bytes, blockStart, blockEnd)
+		require.NoError(t, err, "unexpected error writing req")
+	}
+
+	require.Equal(t, blockStart, uint32(block.meta.StartTime.Unix()))
+	require.Equal(t, blockEnd, uint32(block.meta.EndTime.Unix()))
+
+	// rescan the block and make sure that start/end times are correct
+	blockStart = uint32(time.Now().Add(time.Hour).Unix())
+	blockEnd = uint32(time.Now().Add(2 * time.Hour).Unix())
+
+	blocks, err := wal.RescanBlocks(func([]byte, string) (uint32, uint32, error) {
+		return blockStart, blockEnd, nil
+	}, log.NewNopLogger())
+	require.NoError(t, err, "unexpected error getting blocks")
+	require.Len(t, blocks, 1)
+
+	require.Equal(t, blockStart, uint32(blocks[0].meta.StartTime.Unix()))
+	require.Equal(t, blockEnd, uint32(blocks[0].meta.EndTime.Unix()))
 }
 
 func TestAppendReplayFind(t *testing.T) {
@@ -193,7 +236,7 @@ func testAppendReplayFind(t *testing.T, e backend.Encoding) {
 		require.NoError(t, err)
 		objs = append(objs, bObj)
 
-		err = block.Append(id, bObj)
+		err = block.Append(id, bObj, 0, 0)
 		require.NoError(t, err, "unexpected error writing req")
 	}
 
@@ -211,7 +254,9 @@ func testAppendReplayFind(t *testing.T, e backend.Encoding) {
 	err = appendFile.Close()
 	require.NoError(t, err)
 
-	blocks, err := wal.RescanBlocks(log.NewNopLogger())
+	blocks, err := wal.RescanBlocks(func([]byte, string) (uint32, uint32, error) {
+		return 0, 0, nil
+	}, log.NewNopLogger())
 	require.NoError(t, err, "unexpected error getting blocks")
 	require.Len(t, blocks, 1)
 
@@ -431,7 +476,7 @@ func benchmarkWriteFindReplay(b *testing.B, encoding backend.Encoding) {
 
 		// write
 		for j, obj := range objs {
-			err := block.Append(ids[j], obj)
+			err := block.Append(ids[j], obj, 0, 0)
 			require.NoError(b, err)
 		}
 
@@ -442,7 +487,9 @@ func benchmarkWriteFindReplay(b *testing.B, encoding backend.Encoding) {
 		}
 
 		// replay
-		_, err = wal.RescanBlocks(log.NewNopLogger())
+		_, err = wal.RescanBlocks(func([]byte, string) (uint32, uint32, error) {
+			return 0, 0, nil
+		}, log.NewNopLogger())
 		require.NoError(b, err)
 	}
 }